youtube_backend/requirement.txt
langchain
langchain-core
langchain-community
langchain-openai
youtube-transcript-api
faiss-cpu
tiktoken 
python-dotenv

flask
flask-cors